created CorpusInterface + renamed Corpus -> ArrayCorpus + added methods: getLexemes, setLexemes, updateLexem
